[z0037] Fixed structure CMSG_REQUEST_PARTY_MEMBER_STATS.

(based on commit 9e38c5d)

+++ src/game/GroupHandler.cpp
@@ -778,7 +778,7 @@ void WorldSession::HandleRequestPartyMemberStatsOpcode( WorldPacket &recv_data )
         WorldPacket data(SMSG_PARTY_MEMBER_STATS_FULL, 3+4+2);
         data.appendPackGUID(Guid);
         data << (uint32) GROUP_UPDATE_FLAG_STATUS;
-        data << (uint16) MEMBER_STATUS_OFFLINE;
+        data << (uint8)  MEMBER_STATUS_OFFLINE;
         SendPacket(&data);
         return;
     }
@@ -794,7 +794,7 @@ void WorldSession::HandleRequestPartyMemberStatsOpcode( WorldPacket &recv_data )
 
     Powers powerType = player->getPowerType();
     data << (uint32) mask1;                                 // group update mask
-    data << (uint16) MEMBER_STATUS_ONLINE;                  // member's online status
+    data << (uint8) MEMBER_STATUS_ONLINE;                  // member's online status
     data << (uint16) player->GetHealth();                   // GROUP_UPDATE_FLAG_CUR_HP
     data << (uint16) player->GetMaxHealth();                // GROUP_UPDATE_FLAG_MAX_HP
     data << (uint8)  powerType;                             // GROUP_UPDATE_FLAG_POWER_TYPE
